+++ src/components/molecules/purchasedItem/index.tsx
@@ -1,24 +1,7 @@
 import { getDownloadURL, ref } from 'firebase/storage';
-import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { X } from 'react-feather';
 import { storage } from '../../../../firebase/firebase';
-import { Count } from '../count';
-import {
-  StyledModal,
-  modalStyle,
-  ModalWrapper,
-  ModalCloseButton,
-  ModalLeft,
-  ModalImage,
-  ModalRight,
-  ModalItemName,
-  SalerWrapper,
-  ModalSeller,
-  ModalItemPrice,
-  BuyButtonWrapper,
-  StyledButton,
-} from '../item/style';
 import { Wrapper, StyledImage, Text, Price, ItemInfo, DeleteButton, RightParts } from './style';
 
 type ItemProps = {
@@ -27,12 +10,10 @@ type ItemProps = {
   price: number;
   imageId: string;
   salerName: string;
+  onModalOpen: () => void;
 };
 
-export const PurchasedItem = ({ id, name, price, imageId, salerName }: ItemProps) => {
-  const [modalIsOpen, setIsOpen] = useState(false);
-  const [count, setCount] = useState(1);
-  const router = useRouter();
+export const PurchasedItem = ({ name, price, imageId, onModalOpen }: ItemProps) => {
   const [url, setURL] = useState('');
 
   useEffect(() => {
@@ -45,47 +26,9 @@ export const PurchasedItem = ({ id, name, price, imageId, salerName }: ItemProps
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const onPurchase = (itemId: string, count: number) => {
-    if (window.localStorage) {
-      const defaultItemsJson = localStorage.getItem('purchasedItems');
-      const defaultItems = defaultItemsJson == null ? [] : JSON.parse(defaultItemsJson);
-      const purchasedItems = [...defaultItems, { id: itemId, count: count }];
-      let purchasedItemsJson = JSON.stringify(purchasedItems, undefined, 1);
-      localStorage.setItem('purchasedItems', purchasedItemsJson);
-    }
-  };
-
-  const onModalClose = () => {
-    setCount(1);
-    setIsOpen(false);
-  };
-
   return (
     <>
-      <StyledModal isOpen={modalIsOpen} onRequestClose={() => onModalClose()} style={modalStyle}>
-        <ModalWrapper>
-          <ModalCloseButton onClick={() => setIsOpen(false)}>
-            <X />
-          </ModalCloseButton>
-          <ModalLeft>
-            <ModalImage src={url} />
-          </ModalLeft>
-          <ModalRight>
-            <ModalItemName>{name}</ModalItemName>
-            <SalerWrapper>
-              <ModalSeller onClick={() => router.push(`/list/${salerName}`)}>
-                {salerName}
-              </ModalSeller>
-            </SalerWrapper>
-            <ModalItemPrice>{price * count}</ModalItemPrice>
-            <BuyButtonWrapper>
-              <Count count={count} onChange={setCount} />
-              <StyledButton text='カートに入れる' onClick={() => onPurchase(id, count)} />
-            </BuyButtonWrapper>
-          </ModalRight>
-        </ModalWrapper>
-      </StyledModal>
-      <Wrapper onClick={() => setIsOpen(true)}>
+      <Wrapper onClick={() => onModalOpen()}>
         <StyledImage src={url} />
         <ItemInfo>
           <Text>{name}</Text>
